Investigated Issue #4
There does not appear to be a way to get the bluetooth module into the correct mode. I have an outstanding request at robolink.
I did learn that setting the broadcast mode to mute returns it to the yellow state.
I also learned that making the mode active provides the additional messages I was missing.
Code around connecting has been reworked and should be more robust. More testing with multiple drones is required.